refactor: the paginator refreshes the page whenever a page is changed
Previously, whenever the user clicked on a page number, the data was fetched with ajax requests and the page was not referehed.
Now the page refreshes whenever a user clicks on a page number.
The reason for that is to be able to go to a specific reply from a link. For example, in the thread or categories  index view, if we want to go to the most recent reply directly, we can click on teh link and the page will be refreshed and with the use of the id attribute on each reply we can scroll directly to that reply.