fix: the position of a reply depends now on the replies count column of a thread
The body of a thread is displayed as a reply but it is not.
Threfore, when a tread is created - a reply is created, but in this case it should nout cound as a reply, that's why the replies count is not increased.
However, the position of this reply is the first in the list.