fix: when a user updates the notification settings, the existing setting is updated isteand of creating a new record
users can disable or enable email notifications for a thread subscription.
before:when a user decides to update thread subscription notification settings( disable the email notifications)  a new record was created with the new setting.
now: when a user decides to update thread subscription notification settings( disable the email notifications)  the existing record is updated