fix: fixes an issue where in the foreach only the value was being used instead of both key and values
the foreach was using only the keys of the array and in this case the keys and the value were the other way araound ( the keys were the value sand the value were the keys ).

the getFilters method fetches only the request parameters based on the filters that are supported per class.