refactor: collect the necesscary information for each model when is searched with algolia
When a user sends a saearch query, algolia is responsible for searching the given query and returns an array with the models and the ids of each record that was found from the search.

In order to load the necesscary relationships and additional information such as replies_count, i had to create a new class named GetScoutModels which loads all the information for each model and finally returns the collection to Scout to continue, and replace a line of Scount extended source code. More specifically the ModelsResolver.php