fix: in case the data are retrieved from the activities table, it transforms the data in order to get the subjec of the activity
when the search query is empty, the actiivties table is used to get data from the database.
in that case the subject of the activity is required ( which is a thread,reply, comment, profile_post) in order to display the results.
this commit adds a function that transforms the paginated data to fetch only the subject