Feature a user can now search profile posts from the profile post search tab in combination with filters
A user can now select the profile post search tab and search profile posts and comments. The search can be either typing a search word in combination with filters or just using the filters. The profile post filters are who made the post, the user profile that the post was posted on and the date it was posted.