Refactor use factories to get the desired search strategy and model filter.
based on the user's request a search strategy and a model filter is created. if the user enters a search word then the algolia search strategies are used, whereas if the user does not enter a search word, the databae serach strategies are used. Then for each algolia or database search strategy, based on the other query parameters the final search strategy is created an returned. For the model filter, based on the type of the request a model filter is created.