Refactor reduce the possible search index/strategy classes since algolia returns all records when empty string is being passed to search method
before: when an empty string or no string was passed to the search bar, the application directly queried the database to get all the corresponding records.
now: algolia returns all records when empty string is being passed to the search method. therefore there is no need to get the activity of any post.