Fix strip out html tags and highlight query words when a user searches the forum
the body of a thread/reply might have html tags, however these tags should be visible only when visiting a thread and not when searching results or viewing user's activity. Threfore, in all the other cases ( except when visiting a thread ) the tags are stripped out. Additionally, when a user searches the forum the query words are higlighted ( bold )