Refactor generalize the use of reply and replies components in order to use them also for conversations
before the reply and replies components use the thread and threadPoster variables. however, since the same components can be used for conversations, i renamed the thread and threadPoster to repliable and repliablePoster. Additionally, the original poster label will be displayed only when the reply has a repliable type of thread