Refactor pass the repliable to the reply component instead of the repliable poster
all replies have a repliable but not all repliables have a poster. for example, a thread is a repliable for a reply, and a thread has a poster. however, a conversation is also a repliable for a reply but it doesn't have a repliable poster. therefore i generalize the property to fit all cases