Feature a user can hide or leave a conversation permanently
if a user selects to hide ( leave temporarily a conversation ) then the user will no longer see the conversation in the conversation list but the conversation will reappear once another user adds a new message to the hidden conversation. ( the user will see every new message )

if a user selects to leave permanently the conversation, then the user will no longer see the conversation in the conversation list and the conversation will never reappear if a new message is added to the conversation.