WIP recreate the validation error messages when there is an array attribute
when there is an array attribute in the validation rules, then for each invalid value in the array a new error message will be created. however, what i want is to group the invalid values and their corresponding messages in order to have one unique error message and all the invalid values together. For example 'the following users do not exist: george,orestis,john

instead of having 3 error messages where each errror message has only 1 name that does not exist