Feature: separate files for creaating names array from a string, depending on the case
There are two cases where creating an array of names from a string is needed.

The first case is to get the string (of names) from the session and then replace it with an array. Therefore the array of names is not returned but instead it replaces the request value in the session.

The second case is when using filters where the string with the names is passed as an argument to a method and therefore the array of names is returned after calling the action.