Refactor: use a query scope to determine whether a thread has been updated
before: in orderr to determine whether the thread was updated, an accessor was used which was not an optimal way.
now: a local query scope is used, which adds a column named has_been_updated which determines whether the thread as been updated since the last time it was read. When an authenticated user visits a therad, the thread is marked as read, however for guests a thread is never marked as read.
When a thread is read, the has_been_updated column is false