Refactor: use recursive query to find sub categories
before: there was a relationship between parent categories and threads, and sub-categories and threads. Therefore, in order to count the threads, replies and to find the most recently active thread for a category, depending on whether the category is parent or sub, a different relationship had to be used.

now: with the use of recursive queries which are encapsulated in query scopes, we can count the threads, replies and find the most recently active thread associated with a category, regardless of whether the category is parent or children