Test: change the assertion methods
now the assertion methods for threads, replies, comments and profile posts, expect the array of results and the item that we want assert that is contained in the results.

before: i had to filter and find the a specific result and then compare it with the item that i want.