Refactor: improve encapsulation
the repliable model should not know the internals of the reply. in order to create a reply, the user_id and the body are required, however a thread should not know what are the attributes that are required to create a reply. thus when a repliable creates a new reply, the required attributes are passed as parameter and the reply knows how to associate a user to it, using the setPoster method