Feature : append visitor information on every axios request
on every axios request that is sent by Vue, the backend will append extra data in the response such as, the avatar path, the unviewed notifications count and unread conversations count.
The result of this feature is to simulate real time notifications.
For example, when the user updates the avatar, an axios request is sent to laravel, if in the meantime the user has a new notification but hasn't refresh the page yet, laravel will append the unviewed noticiations count and as a result, the user will see that there is a new notification without actually refreshing the page.