Fix: issue where null was returned for attribute conversation admin
before: the conversation admin attribute was being added within the participants relationship. The problem was that i was using eager load to get the participants of a conversation together with the converastion_admin attribute, but it seems that when the converastion wants to eager load the participants relationship, it adds the where conversation.id = ?  at the end, and the conversationAdmin scope didn't have yet the conversation id in the query.

now: the participants are fetched separately