Refactor: rename followedBy to followers and follows to followings
the name of the relationship to get the followers of a user was followedBy whereas now it is followers. the relationship to get the users that a user follows was follows and now it is followings. According to the new names of the relationships all the proper changes have been made to remain consistent