Fix: issue where filters were ignored
when you go to the profile posts index page, you can see all profile posts and then you can filter those posts. the issue was that after a filter was selected, when users change page, that filter was not taken into account. therefore, by enabling the query string when paginating, the paginator now includes the current query string parameters ( which are basically the fitlers )