Test: ensure that statistics are stored in cache
the fake statistics tests ensure that the statistics methods in the app ( i.e increment() ) are called properly. The fake statistics classes are used to avoid using the actual cache when running tests. The real statistics classes, are used to test that the application actually stores the statistics data in cache