Performance: reduce the number of queries
before: the notify thread subscribers listener was using the HandlesNotifications trait which makes a query for every user to check whether is ignore or not.  Also, the subscription for the even thread is eager loaded for all subscribers, which also reduces a query to the database, becaus the ThreadHasNewReply notification, needs the subscription record
now: the notIgnoring scope filters out the subsribers that are ignoring the authenticated user.