Fix: issue where the id of conversations was wrong
when the withHasBeenUpdated is called, there is a conflict with the columns that are fetched from the db, as a result, the id of conversations that was returned, was wrong while the other columns were correct. That caused issue with the relationships that are eager loaded, since wrong ids were fetched for the conversations, wrong relationships or even no relationships at all were eager loaded.